requirement.txt
openai
dotenv
pandas
openpyxl
google-generativeai
click